+++ src/Angor/Avalonia/Angor.Contexts.Funding.Tests/NostrDecrypterTests.cs
@@ -16,7 +16,7 @@ public class NostrDecrypterTests(ITestOutputHelper output)
     {
         // Arrange
         var sut = CreateSut();
-        var nostrMessage = new InvestmentMessage("SOME ID", "c57cce49441740454e54fee15131b9699d54d4928a5eaa0756214dc21961cf1b", "gYpNKJuzweBOfsOXOOInoyEXI+jlOxs2dQ//Yk3CSpkvUQPs6od9eq/HCUGBPetHlTBlmlJheVNYzl+u+4NcShzgV2kmOb7rbH88UUCZXRyXdxwwnGU8rvM4BKNc/WqS3Rsyny+hDpy/YpEpsSMH46kTAqPm6+PjyyIoTGNMqAui95RdTNMHXju6hLXse6kTL93aD8xpCsLVZ773PaJxtO8qzVmCJ9Ton+TMkD3ujxcUAv/Dy03vRQWp96iNSfsBif6OUlenGPk5YUzgKscjNVVxOgLCB7hdlM9NQiTDB3EKabl6p7L9Dg77UftRcrHwYE44QYtgOjTggNRtoWYhAxgITp4IHtIOLJRIBexlxrY+2R8Zfkd1w0qvTm41zi9eITB0/rJEE/u8jMFJLVxDMLcaGVWuj1/9NlcZ9jTpf5ANOr6VnWDi2JRUsfJXDxUNyYSQttoclZVYD1gLjv/+ags+aBQzpiz5XOAC1NVOouKTE4Wj4b9MLvMJLMfyjtmBPBc++7uy0dOX6YVQqorOzNtk6NkOQFPDdCZK5zmQB2twWwwo6TlzKUT3n3msoeHi7AuBLAxoD7v2hmY61SMt7Z8tM4kOmaxMF+ed1ZKrD1gyfWniOD/4GXj+lovTPw4vOZRLmfYTTCKJROB/qBheH5tyuNahn4FiIWYxATTBO3jJoKw9evQdf8haDptwlPTZUwlf3C+T0nf+9Z+bXSWaRPsORQzNgWYnRys/G4OpodXiVFwdH+6rYeXDMXAiCMEDXBBfT2e5QZxVuJuSbhTW75jNvDEaHrg12UxvA+7nqCy4rpHC3nDQaABEprXdYK6bANqIWSsNCsmUhSjLfpH+gljJ5iiH+yE6kf/tGGNnSS4pKa9cPSMbSB68eIWwUeaNRRBtcJkvr42K35jUE2vEQhqs3ojQlMItgy4/YvcXP6+DMzGZbxdDWO8u0xsE0qNpkCavwi/+UgVsYj5dEK32nt8kiBryr6/UJyWprOa8lUQ=?iv=+uu8GD5eAnBhjP3FjCVeSQ==", DateTime.Now);
+        var nostrMessage = new DirectMessage("SOME ID", "c57cce49441740454e54fee15131b9699d54d4928a5eaa0756214dc21961cf1b", "gYpNKJuzweBOfsOXOOInoyEXI+jlOxs2dQ//Yk3CSpkvUQPs6od9eq/HCUGBPetHlTBlmlJheVNYzl+u+4NcShzgV2kmOb7rbH88UUCZXRyXdxwwnGU8rvM4BKNc/WqS3Rsyny+hDpy/YpEpsSMH46kTAqPm6+PjyyIoTGNMqAui95RdTNMHXju6hLXse6kTL93aD8xpCsLVZ773PaJxtO8qzVmCJ9Ton+TMkD3ujxcUAv/Dy03vRQWp96iNSfsBif6OUlenGPk5YUzgKscjNVVxOgLCB7hdlM9NQiTDB3EKabl6p7L9Dg77UftRcrHwYE44QYtgOjTggNRtoWYhAxgITp4IHtIOLJRIBexlxrY+2R8Zfkd1w0qvTm41zi9eITB0/rJEE/u8jMFJLVxDMLcaGVWuj1/9NlcZ9jTpf5ANOr6VnWDi2JRUsfJXDxUNyYSQttoclZVYD1gLjv/+ags+aBQzpiz5XOAC1NVOouKTE4Wj4b9MLvMJLMfyjtmBPBc++7uy0dOX6YVQqorOzNtk6NkOQFPDdCZK5zmQB2twWwwo6TlzKUT3n3msoeHi7AuBLAxoD7v2hmY61SMt7Z8tM4kOmaxMF+ed1ZKrD1gyfWniOD/4GXj+lovTPw4vOZRLmfYTTCKJROB/qBheH5tyuNahn4FiIWYxATTBO3jJoKw9evQdf8haDptwlPTZUwlf3C+T0nf+9Z+bXSWaRPsORQzNgWYnRys/G4OpodXiVFwdH+6rYeXDMXAiCMEDXBBfT2e5QZxVuJuSbhTW75jNvDEaHrg12UxvA+7nqCy4rpHC3nDQaABEprXdYK6bANqIWSsNCsmUhSjLfpH+gljJ5iiH+yE6kf/tGGNnSS4pKa9cPSMbSB68eIWwUeaNRRBtcJkvr42K35jUE2vEQhqs3ojQlMItgy4/YvcXP6+DMzGZbxdDWO8u0xsE0qNpkCavwi/+UgVsYj5dEK32nt8kiBryr6/UJyWprOa8lUQ=?iv=+uu8GD5eAnBhjP3FjCVeSQ==", DateTime.Now);
         
         // Act
         var clientKeyResult = await sut.Decrypt(Guid.NewGuid(), new ProjectId("angor1qatlv9htzte8vtddgyxpgt78ruyzaj57n4l7k46"), nostrMessage);
