+++ src/Angor/Avalonia/Angor.Contexts.Funding/Shared/DirectMessage.cs
@@ -1,0 +1,3 @@
+namespace Angor.Contexts.Funding.Shared;
+
+public record DirectMessage(string Id, string InvestorNostrPubKey, string Content, DateTime Created);
