+++ src/Angor/Avalonia/Angor.Contexts.Funding/Shared/INostrDecrypter.cs
@@ -1,11 +1,9 @@
-using Angor.Contexts.Funding.Founder.Operations;
 using Angor.Contexts.Funding.Projects.Domain;
-using Angor.Contexts.Funding.Shared;
 using CSharpFunctionalExtensions;
 
-namespace Angor.Contexts.Funding;
+namespace Angor.Contexts.Funding.Shared;
 
 public interface INostrDecrypter
 {
-    Task<Result<string>> Decrypt(Guid walletId, ProjectId projectId, InvestmentMessage nostrMessage);
+    Task<Result<string>> Decrypt(Guid walletId, ProjectId projectId, DirectMessage nostrMessage);
 }
