--- src/Angor/Avalonia/Angor.Contexts.Funding/Shared/InvestmentMessage.cs
@@ -1,3 +1,0 @@
-namespace Angor.Contexts.Funding.Shared;
-
-public record InvestmentMessage(string Id, string InvestorNostrPubKey, string Content, DateTime Created);
